refactor: extract MonsterTemplateCard and MonsterTemplateEditor from page.tsx
- Extract MonsterTemplateCard → app/monsters/MonsterTemplateCard.tsx
- Extract MonsterTemplateEditor → app/monsters/MonsterTemplateEditor.tsx
- Rename inline normalizeSpeed helper to formatSpeedValue (avoids confusion with lib/import/transformMonster.ts:normalizeSpeed)
- Add htmlFor/id accessibility attributes to MonsterTemplateEditor inputs
- Move tests/unit/monstersPage.test.tsx → tests/unit/components/MonstersPage.test.tsx
- Add tests/unit/components/MonsterTemplateEditor.test.tsx (18 tests)

Closes #378
