+++ app/monsters/page.tsx
@@ -3,12 +3,11 @@
 import { useState, useEffect, useMemo, useCallback } from 'react';
 import Link from 'next/link';
 import { ProtectedRoute } from '@/lib/components/ProtectedRoute';
-import { CreatureStatBlock } from '@/lib/components/CreatureStatBlock';
-import { CreatureStatsForm } from '@/lib/components/CreatureStatsForm';
-import { AlignmentSelect } from '@/lib/components/AlignmentSelect';
-import { MonsterTemplate, normalizeAlignment } from '@/lib/types';
+import type { MonsterTemplate } from '@/lib/types';
 import { GLOBAL_USER_ID } from '@/lib/constants';
 import { filterMonsters, getAvailableTypes } from './filterUtils';
+import { MonsterTemplateCard } from './MonsterTemplateCard';
+import { MonsterTemplateEditor } from './MonsterTemplateEditor';
 
 export function MonstersContent() {
   const [userTemplates, setUserTemplates] = useState<MonsterTemplate[]>([]);
@@ -347,297 +346,6 @@ export function MonstersContent() {
           </div>
         </div>
       )}
-    </div>
-  );
-}
-
-function MonsterTemplateCard({
-  template,
-  isGlobal,
-  canEdit = true,
-  onEdit,
-  onDelete,
-  onCopy,
-  isCopying = false,
-}: {
-  template: MonsterTemplate;
-  isGlobal: boolean;
-  canEdit?: boolean;
-  onEdit: () => void;
-  onDelete: () => void;
-  onCopy?: () => void;
-  isCopying?: boolean;
-}) {
-  return (
-    <div className={`rounded-lg p-4 ${isGlobal ? 'bg-gray-800 border border-purple-600' : 'bg-gray-800'}`}>
-      <div className="flex justify-between items-start mb-4">
-        <div>
-          <div className="flex items-center gap-2">
-            <h3 className="text-xl font-semibold">{template.name}</h3>
-            {isGlobal && <span className="px-2 py-1 bg-purple-600 text-xs rounded">Global</span>}
-          </div>
-          {(template.size || template.type) && (
-            <p className="text-sm text-gray-400 mt-1">
-              {template.size} {template.type}
-              {template.challengeRating !== undefined && ` (CR ${template.challengeRating})`}
-            </p>
-          )}
-        </div>
-        <div className="flex gap-2">
-          {isGlobal && onCopy && (
-            <button
-              onClick={onCopy}
-              disabled={isCopying}
-              className="bg-green-600 hover:bg-green-700 disabled:bg-gray-600 disabled:cursor-not-allowed px-3 py-1 rounded text-sm"
-              aria-label={`Copy ${template.name} to your library`}
-            >
-              {isCopying ? 'Copying...' : 'Copy'}
-            </button>
-          )}
-          {canEdit && (
-            <>
-              <button
-                onClick={onEdit}
-                className="bg-blue-600 hover:bg-blue-700 px-3 py-1 rounded text-sm"
-              >
-                Edit
-              </button>
-              <button
-                onClick={onDelete}
-                className="bg-red-600 hover:bg-red-700 px-3 py-1 rounded text-sm"
-              >
-                Delete
-              </button>
-            </>
-          )}
-        </div>
-      </div>
-      <CreatureStatBlock
-        abilityScores={template.abilityScores}
-        ac={template.ac}
-        acNote={template.acNote}
-        hp={template.hp}
-        maxHp={template.maxHp}
-        skills={template.skills}
-        savingThrows={template.savingThrows}
-        damageResistances={template.damageResistances}
-        damageImmunities={template.damageImmunities}
-        damageVulnerabilities={template.damageVulnerabilities}
-        conditionImmunities={template.conditionImmunities}
-        senses={template.senses}
-        languages={template.languages}
-        traits={template.traits}
-        actions={template.actions}
-        bonusActions={template.bonusActions}
-        reactions={template.reactions}
-        isCompact={true}
-      />
-    </div>
-  );
-}
-
-function MonsterTemplateEditor({
-  template,
-  onSave,
-  onCancel,
-  isNew,
-  isGlobal,
-}: {
-  template: MonsterTemplate;
-  onSave: (template: MonsterTemplate) => void;
-  onCancel: () => void;
-  isNew: boolean;
-  isGlobal: boolean;
-}) {
-  // Helper function to convert speed object to human-readable string
-  const normalizeSpeed = (speedValue: unknown): string => {
-    if (typeof speedValue === 'string') {
-      return speedValue;
-    }
-    if (typeof speedValue === 'object' && speedValue !== null) {
-      return Object.entries(speedValue as Record<string, string>)
-        .map(([key, value]) => `${key} ${value}`)
-        .join(', ');
-    }
-    return '30 ft.';
-  };
-
-  const [name, setName] = useState(template.name);
-  const [size, setSize] = useState(template.size);
-  const [type, setType] = useState(template.type);
-  const [alignment, setAlignment] = useState(
-    normalizeAlignment(template.alignment) ?? '',
-  );
-  const [speed, setSpeed] = useState(normalizeSpeed(template.speed));
-  const [challengeRating, setChallengeRating] = useState(template.challengeRating);
-  const [source, setSource] = useState(template.source || '');
-  const [description, setDescription] = useState(template.description || '');
-  const [stats, setStats] = useState(template);
-  const [saving, setSaving] = useState(false);
-  const [validationError, setValidationError] = useState<string | null>(null);
-
-  const handleSave = async () => {
-    setValidationError(null);
-
-    if (!name.trim()) {
-      setValidationError('Monster name is required');
-      return;
-    }
-
-    if (stats.hp > stats.maxHp) {
-      setValidationError('Current HP cannot be greater than Max HP');
-      return;
-    }
-
-    setSaving(true);
-    try {
-      const updatedTemplate: MonsterTemplate = {
-        ...stats,
-        name,
-        size,
-        type,
-        alignment: normalizeAlignment(alignment),
-        speed,
-        challengeRating,
-        source: source || undefined,
-        description: description || undefined,
-        updatedAt: new Date(),
-      };
-      await onSave(updatedTemplate);
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  const title = isGlobal ? 'Global Monster' : 'Personal Monster';
-
-  return (
-    <div className={`rounded-lg p-6 mb-6 border-2 ${isGlobal ? 'border-purple-500 bg-gray-800' : 'border-blue-500 bg-gray-800'}`}>
-      <h3 className="text-2xl font-bold mb-4">{isNew ? `Create ${title}` : `Edit ${title}`}</h3>
-
-      {validationError && (
-        <div className="p-3 bg-red-900 border border-red-700 rounded text-red-200 mb-4">
-          {validationError}
-        </div>
-      )}
-
-      {/* Basic Monster Info */}
-      <div className="mb-6 pb-6 border-b border-gray-700">
-        <h4 className="font-bold text-gray-300 mb-4">Basic Info</h4>
-        <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
-          <div>
-            <label className="block mb-1 text-sm font-bold">Name</label>
-            <input
-              type="text"
-              value={name}
-              onChange={e => setName(e.target.value)}
-              className="w-full bg-gray-700 rounded px-3 py-2 text-white"
-              disabled={saving}
-            />
-          </div>
-
-          <div>
-            <label className="block mb-1 text-sm font-bold">Size</label>
-            <select
-              value={size}
-              onChange={e => setSize(e.target.value as any)}
-              className="w-full bg-gray-700 rounded px-3 py-2 text-white"
-              disabled={saving}
-            >
-              <option value="tiny">Tiny</option>
-              <option value="small">Small</option>
-              <option value="medium">Medium</option>
-              <option value="large">Large</option>
-              <option value="huge">Huge</option>
-              <option value="gargantuan">Gargantuan</option>
-            </select>
-          </div>
-
-          <div>
-            <label className="block mb-1 text-sm font-bold">Type</label>
-            <input
-              type="text"
-              value={type}
-              onChange={e => setType(e.target.value)}
-              placeholder="e.g., humanoid, beast"
-              className="w-full bg-gray-700 rounded px-3 py-2 text-white"
-              disabled={saving}
-            />
-          </div>
-
-          <div>
-            <AlignmentSelect value={alignment} onChange={setAlignment} disabled={saving} showExtendedAlignments />
-          </div>
-
-          <div>
-            <label className="block mb-1 text-sm font-bold">Speed</label>
-            <input
-              type="text"
-              value={speed}
-              onChange={e => setSpeed(e.target.value)}
-              placeholder="e.g., 30 ft., fly 60 ft."
-              className="w-full bg-gray-700 rounded px-3 py-2 text-white"
-              disabled={saving}
-            />
-          </div>
-
-          <div>
-            <label className="block mb-1 text-sm font-bold">Challenge Rating</label>
-            <input
-              type="number"
-              value={challengeRating}
-              onChange={e => setChallengeRating(parseFloat(e.target.value) || 0)}
-              className="w-full bg-gray-700 rounded px-3 py-2 text-white"
-              disabled={saving}
-              step="0.125"
-              min="0"
-            />
-          </div>
-
-          <div className="md:col-span-2">
-            <label className="block mb-1 text-sm font-bold">Source</label>
-            <input
-              type="text"
-              value={source}
-              onChange={e => setSource(e.target.value)}
-              placeholder="e.g., Monster Manual"
-              className="w-full bg-gray-700 rounded px-3 py-2 text-white"
-              disabled={saving}
-            />
-          </div>
-
-          <div className="md:col-span-3">
-            <label className="block mb-1 text-sm font-bold">Description / Notes</label>
-            <textarea
-              value={description}
-              onChange={e => setDescription(e.target.value)}
-              className="w-full bg-gray-700 rounded px-3 py-2 text-white"
-              disabled={saving}
-              rows={2}
-            />
-          </div>
-        </div>
-      </div>
-
-      {/* Creature Stats */}
-      <CreatureStatsForm stats={stats} onChange={(updatedStats) => setStats(prev => ({ ...prev, ...updatedStats }))} />
-
-      <div className="flex gap-2 mt-6">
-        <button
-          onClick={handleSave}
-          disabled={saving || !name.trim()}
-          className={`hover:opacity-80 disabled:opacity-50 px-4 py-2 rounded ${isGlobal ? 'bg-purple-600' : 'bg-green-600'}`}
-        >
-          {saving ? 'Saving...' : `Save ${title}`}
-        </button>
-        <button
-          onClick={onCancel}
-          disabled={saving}
-          className="bg-gray-600 hover:bg-gray-700 disabled:bg-gray-700 px-4 py-2 rounded"
-        >
-          Cancel
-        </button>
-      </div>
     </div>
   );
 }
